pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kamegrueon <niwaniwaniwaniwatori.gairuze@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.7"
google-api-python-client = "^2.86.0"
pytz = "^2023.3"
google-cloud-storage = "^2.8.0"
google-cloud-bigquery = "^3.10.0"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.1" }
gunicorn = "^21.2.0"
google-cloud-secret-manager = "^2.18.2"
google-cloud-pubsub = "^2.19.7"
pydantic = "^2.6.3"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
pep8-naming = "^0.13.3"
ruff = "^0.6.1"


[tool.poetry.group.types.dependencies]
types-pytz = "^2023.3.0.0"

[tool.ruff]
# 1行の最大文字数
line-length = 119

[tool.ruff.lint]
# チェックするエラーの種類
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
# 除外するエラーの種類
ignore = [
    "I001",  # import library formatting
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904",  # raise ... from ... になっていない場合のエラーを無視
]

# ファイルごとのエラー除外
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

# isort の設定
[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.lint.pyupgrade]
# Python3.8互換のための設定
keep-runtime-typing = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
